# docs/openapi.yaml
openapi: 3.0.3
info:
  title: India In-Time API
  version: "3.6.0"
  description: |
    Production API contract for India In-Time. Request and response schemas are
    intentionally explicit so frontend, mobile, QA, and third-party clients can
    validate payloads without reverse-engineering route implementations.
servers:
  - url: /api
    description: Deployment-relative API root
...
paths:
...
  /time-intelligence/status:
    post:
      tags: [Time Intelligence]
      summary: Score place status for a planned arrival time
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TimeIntelligenceStatusRequest' }
      responses:
        '200':
          description: Per-place status
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/TimeIntelligenceStatus' }
                required: [results]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /time-intelligence/recommend:
    post:
      tags: [Time Intelligence]
      summary: Rank places for a day
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TimeIntelligenceRecommendRequest' }
      responses:
        '200':
          description: Ranked recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  places:
                    type: array
                    items: { $ref: '#/components/schemas/RecommendedPlace' }
                required: [places]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /time-intelligence/day-plan:
    post:
      tags: [Time Intelligence]
      summary: Build a timed day plan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DayPlanRequest' }
      responses:
        '200':
          description: Planned itinerary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayPlanResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /time-intelligence/temporal-profile:
    post:
      tags: [Time Intelligence]
      summary: Build a future time-dependent experience profile for a place
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TemporalProfileRequest' }
      responses:
        '200':
          description: Temporal experience windows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemporalProfileResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /time-intelligence/optimize:
    post:
      tags: [Time Intelligence]
      summary: Optimize an itinerary using projected-arrival GeoAI
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GeoAIOptimizeRequest' }
      responses:
        '200':
          description: Geo-temporally optimized itinerary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeoAIOptimizeResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /time-intelligence/replan:
    post:
      tags: [Time Intelligence]
      summary: Replan the remaining itinerary from current state
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GeoAIReplanRequest' }
      responses:
        '200':
          description: Updated itinerary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeoAIOptimizeResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    Coordinates:
      type: array
      minItems: 2
      maxItems: 2
      items: { type: number }
      example: [17.6868, 83.2185]
...
    Place:
      type: object
      required: [id, name, category, coords]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        coords: { $ref: '#/components/schemas/Coordinates' }
        visitTimeMinutes: { type: integer, minimum: 0 }
        openTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        closeTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        rating: { type: number, minimum: 0, maximum: 5 }
        importance: { type: string }
...
    TimeIntelligenceStatusRequest:
      type: object
      required: [places]
      properties:
        city: { type: string, maxLength: 100 }
        arrivalMinutes: { type: integer, minimum: 0, maximum: 1439, default: 600 }
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        places:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: '#/components/schemas/Place' }
    TimeIntelligenceStatus:
      type: object
      required: [placeId, status]
      properties:
        placeId: { type: string }
        status: { type: string }
        score: { type: number }
        crowdLevel: { type: string }
        weatherImpact: { type: number }
        confidence: { type: number, minimum: 0, maximum: 1 }
        reason: { type: string }
    TimeIntelligenceRecommendRequest:
      type: object
      required: [places]
      properties:
        city: { type: string, maxLength: 100 }
        places:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: '#/components/schemas/Place' }
        startMinutes: { type: integer, minimum: 0, maximum: 1439, default: 540 }
        durationMinutes: { type: integer, minimum: 30, maximum: 1440, default: 480 }
        liveRouting: { type: boolean, default: false }
        lat: { type: number, minimum: -90, maximum: 90 }
        lon: { type: number, minimum: -180, maximum: 180 }
    RecommendedPlace:
      allOf:
        - $ref: '#/components/schemas/Place'
        - type: object
          properties:
            score: { type: number }
            confidence: { type: number, minimum: 0, maximum: 1 }
            why: { type: string }
            etaMinutes: { type: integer, minimum: 0 }
    DayPlanRequest:
      allOf:
        - $ref: '#/components/schemas/TimeIntelligenceRecommendRequest'
      properties:
        maxStops: { type: integer, minimum: 1, maximum: 20, default: 8 }
        includeMeals: { type: boolean, default: true }
    DayPlanResponse:
      type: object
      required: [stops]
      properties:
        stops:
          type: array
          items: { $ref: '#/components/schemas/RecommendedPlace' }
        totalTravelMinutes: { type: integer, minimum: 0 }
        totalVisitMinutes: { type: integer, minimum: 0 }
        finishMinutes: { type: integer, minimum: 0, maximum: 1439 }
    TemporalProfileRequest:
      type: object
      required: [place]
      properties:
        place: { $ref: '#/components/schemas/Place' }
        weather: { $ref: '#/components/schemas/WeatherInput' }
        at: { type: string, format: date-time }
        stepMin: { type: integer, minimum: 15, maximum: 120, default: 30 }
        horizonHours: { type: integer, minimum: 1, maximum: 36, default: 18 }
        personas: { type: array, items: { type: string, maxLength: 40 }, maxItems: 10 }
        tripMode: { type: string, maxLength: 40 }
    TemporalProfileResponse:
      type: object
      properties:
        place: { type: string }
        resolutionMinutes: { type: integer }
        horizonHours: { type: integer }
        windows: { type: array, items: { $ref: '#/components/schemas/ExperienceWindow' } }
        modes: { type: object, additionalProperties: { $ref: '#/components/schemas/ExperienceWindow' } }
        bestWindow: { $ref: '#/components/schemas/ExperienceWindow' }
        confidence: { type: number, minimum: 0, maximum: 100 }
    ExperienceWindow:
      type: object
      properties:
        start: { type: string, pattern: '^\d{2}:\d{2}$' }
        end: { type: string, pattern: '^\d{2}:\d{2}$' }
        peakTime: { type: string, pattern: '^\d{2}:\d{2}$' }
        score: { type: number, minimum: 0, maximum: 100 }
        confidence: { type: number, minimum: 0, maximum: 100 }
        reasons: { type: array, items: { type: string } }
    GeoAIOptimizeRequest:
      type: object
      required: [places]
      properties:
        places: { type: array, minItems: 1, maxItems: 50, items: { $ref: '#/components/schemas/Place' } }
        weather: { $ref: '#/components/schemas/WeatherInput' }
        at: { type: string, format: date-time }
        fromCoords: { type: array, minItems: 2, maxItems: 2, items: { type: number } }
        startMin: { type: integer, minimum: 0, maximum: 1439 }
        endMin: { type: integer, minimum: 1, maximum: 1439 }
        maxStops: { type: integer, minimum: 1, maximum: 12, default: 8 }
        bufferMin: { type: integer, minimum: 0, maximum: 120, default: 15 }
        beamWidth: { type: integer, minimum: 3, maximum: 12, default: 8 }
        personas: { type: array, items: { type: string, maxLength: 40 }, maxItems: 10 }
        tripMode: { type: string, maxLength: 40 }
    GeoAIReplanRequest:
      allOf:
        - $ref: '#/components/schemas/GeoAIOptimizeRequest'
      properties:
        trigger: { type: string, maxLength: 40 }
        reason: { type: string, maxLength: 300 }
    GeoAIOptimizeResponse:
      type: object
      required: [algorithm, stops]
      properties:
        algorithm: { type: string }
        objective: { type: string }
        stopCount: { type: integer, minimum: 0 }
        totalScore: { type: number }
        totalTravelMinutes: { type: integer, minimum: 0 }
        temporalResolutionMinutes: { type: integer }
        stops:
          type: array
          items:
            $ref: '#/components/schemas/OptimizedStop'
        warnings: { type: array, items: { type: string } }
    OptimizedStop:
      allOf:
        - $ref: '#/components/schemas/Place'
        - type: object
          properties:
            departAt: { type: string }
            arriveAt: { type: string }
            leaveAt: { type: string }
            rawArrivalAt: { type: string }
            waitingMinutes: { type: integer, minimum: 0 }
            stayMinutes: { type: integer, minimum: 0 }
            travelMinutes: { type: integer, minimum: 0 }
            travelSource: { type: string }
            timingFit: { type: number, minimum: 0, maximum: 100 }
            optimizationScore: { type: number }
            bestWindow: { $ref: '#/components/schemas/ExperienceWindow' }
    WeatherInput:
      type: object
      additionalProperties: true
      description: Current/future weather data supplied by the weather provider.
...